+++ lib/sheet/modal_base.dart
@@ -3,7 +3,7 @@ part of 'sheet.dart';
 /// A callback invoked to notify or request a change in drawer or sheet size.
 typedef ChangeDrawerSize = void Function({required bool isExpanded});
 
-abstract class ModalBase<T extends ModalBaseParams> {
+abstract base class ModalBase<T extends ModalBaseParams> {
   const ModalBase({required this.params});
 
   final T params;